Bookstore-website/backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
requests==2.25.1
virtualenv==20.25.1